Extract AuthMethod class as independent class file
